chore: Moved accept/reject outgoing requests to socket.on, moved friend components to separate folder
