+++ frontend/src/app/page.tsx
@@ -1,4 +1,4 @@
-import FriendList from '@/components/FriendList';
+import FriendList from '@/components/friend/FriendList';
 import ChannelHeader from '@/components/ChannelHeader';
 import { ChannelList } from '@/components/ChannelList';
 export default function Home() {
