+++ frontend/src/components/friend/FriendDisplay.tsx
